Moving check-for-update in its own class
